% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply a filter constructed with \link{build_filter}}
\usage{
apply_filter(filt, x, na.skip = TRUE)
}
\arguments{
\item{filt}{Filter as returned from \code{\link[=build_filter]{build_filter()}} or \code{\link[gsignal:butter]{gsignal::butter()}}}

\item{x}{Vector of data to filter}

\item{na.skip}{(TRUE or FALSE) to skip NAs in the data set.}
}
\value{
Filtered data set
}
\description{
Wrapper function for \code{gsignal::filtfilt} to apply a filter to a dataset.
Potentially skips NAs in the data set (see \code{\link[=skip_na]{skip_na()}} for details).
}
