% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_filter}
\alias{build_filter}
\title{Constructs a smoothing filter}
\usage{
build_filter(lo = NULL, hi = NULL, sampfreq, order = 13)
}
\arguments{
\item{lo}{Lower frequency cutoff in Hz. If only \code{lo} is specified, then
\code{build_filter} constructs a high pass filter}

\item{hi}{Upper frequency cutoff in Hz. If only \code{hi} is specified, then
\code{build_filter} constructs a high pass filter}

\item{sampfreq}{Sampling frequency for the data in Hz.}

\item{order}{(optional) Order for the filter}
}
\value{
Filter parameters in Sos form
}
\description{
Wrapper function for \code{gsignal::butter} to construct a bandpass (or
low or high pass) filter at a particular frequency. Uses the Sos form for
the filter, which works better numerically particularly for very low
frequencies.
}
\examples{
# Low pass filter with a cutoff at 0.5Hz for data sampled at 100Hz
lopass <- build_filter(lo=0.5, sampfreq=100)

# Band pass filter that passes frequencies between 0.5 and 10Hz
bandpass <- build_filter(lo=0.5, hi=10, sampfreq=100)
}
