% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.out}
\alias{check.out}
\title{Helper function to check the \code{.out} and \code{.out_default} variables used in
this package}
\usage{
check.out(.data, .out, .out_default, overwrite)
}
\arguments{
\item{.data}{Data frame}

\item{.out}{Name for output columns}

\item{.out_default}{Named list containing default names for output columns}

\item{overwrite}{TRUE or FALSE to overwrite columns.}
}
\value{
The updated \code{.out} list.
}
\description{
\itemize{
\item If \code{.out} is NULL, returns \code{.out_default}.
\item If \code{.out} is a vector containing strings, checks to make sure that there
are the same number of strings in \code{.out} as items in \code{.out_default}
\item If \code{.out} is a list with named elements, checks to make sure the names
are present in \code{.out_default}. Ignores names not present in \code{.out_default},
but gives a warning. If any elements in \code{.out_default} are not in \code{.out},
uses the values from \code{.out_default}. Sorts the items in the same order
as in \code{.out_default}.
}
}
\details{
Checks if there are columns with the same names in the data frame \code{.data}.
Gives a warning if there are such columns and \code{overwrite} is TRUE, and an
error if \code{overwrite} is FALSE.
}
