% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_overwrite_columns}
\alias{check_if_overwrite_columns}
\title{Check if a data frame has columns that we might overwrite}
\usage{
check_if_overwrite_columns(df, newcols, overwrite)
}
\arguments{
\item{df}{Data frame}

\item{newcols}{Names of the columns, as strings}

\item{overwrite}{TRUE or FALSE to overwrite the columns}
}
\description{
Helper function for other functions in the package that create new columns
in a data frame, to check if the columns are already present in the data
frame.
}
\details{
Produces a warning if the columns are present but \code{overwrite} is true, and
an error if \code{overwrite} is false.
}
\examples{
df <- data.frame(a=c(1,2,3), b=c(1,2,3))

# this should give a warning
check_if_overwrite_columns(df, c('a', 'd'), overwrite=TRUE)
}
