% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Estimates curvature for a single curve}
\usage{
curvature(s, x, y, method = "angle")
}
\arguments{
\item{s}{Arc length along the body.}

\item{x, y}{Coordinates of each point along the body}

\item{method}{("xy" or "angle") for the direct formula or for the angle
derivative}
}
\value{
Curvature.
}
\description{
Estimates curvature either directly through derivatives of the x and y
coordinates relative to arc length, or as the derivative of segment angle
relative to arc length.
}
\details{
Assumes that points are in order along the curve from head to tail.
}
\concept{pipeline}
