% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{deriv}
\alias{deriv}
\title{Estimate first or second derivatives for dy/dx.}
\usage{
deriv(x, y, ord = 1, method = "direct", ends = "forwardback")
}
\arguments{
\item{x}{x variable. Does not need to be evenly spaced.}

\item{y}{y variable.}

\item{ord}{Order of the derivative (1 or 2).}

\item{method}{Method for taking second derivatives. Either
\itemize{
\item 'direct' (default) Uses a direct formula, based on a central difference of
forward and backward differences, from \link{https://mathformeremortals.wordpress.com/2013/01/12/a-numerical-second-derivative-from-three-points/}
\item 'repeat' Repeat two first derivatives.
}}
}
\value{
Derivative of y relative to x.
}
\description{
Uses central differencing where possible.
}
