% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{find_gaps_df}
\alias{find_gaps_df}
\title{Find gaps in a data series}
\usage{
find_gaps_df(.data, cols, .frame = frame, .out = c("gaplen"))
}
\arguments{
\item{.data}{Data frame containing the midlines.}

\item{cols}{Columns containing the components to be smoothed. Often these will
be the x and y coordinates of the midline.}

\item{.frame}{Column identifying frames (default is \code{frame})}

\item{.out}{Name of the column to contain the length of the gap.}
}
\value{
The data frame with a new column.
}
\description{
Looks for NAs (gaps) in a data series and counts the number of NAs.
Useful for \code{smooth_points_df}, which can fill gaps up to a certain length.
0 means no gap, 1 means a single NA, and so forth.
}
