% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmechr-package.R
\docType{package}
\name{fishmechr-package}
\alias{fishmechr}
\alias{fishmechr-package}
\title{fishmechr: Computes Kinematic Parameters for Swimming}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tytell.github.io/fishmechr/}
}

}
\author{
\strong{Maintainer}: Eric Tytell \email{eric.tytell@tufts.edu} (\href{https://orcid.org/0000-0002-6603-9448}{ORCID})

}
\keyword{internal}
