% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishwidth}
\alias{fishwidth}
\title{Fish body width profiles}
\format{
A data frame with 20 rows and 3 columns:
\describe{
\item{s}{Arc length as a fraction of body length (0 = head, 1 = tail)}
\item{eelwidth}{Body width of the eel as a fraction of body length}
\item{ammowidth}{Body width of the lamprey as a fraction of body length}
}
}
\usage{
fishwidth
}
\description{
Body width (diameter) as a function of fractional arc length for an eel
(\emph{Anguilla rostrata}) and a sea lamprey (\emph{Petromyzon marinus}). Used for
estimating the center of mass via \code{\link[=get_midline_center_df]{get_midline_center_df()}}.
}
\keyword{datasets}
