% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_frequency.R
\name{get_body_cycle_numbers_df}
\alias{get_body_cycle_numbers_df}
\title{Gets oscillation cycle numbers for a midline data set}
\usage{
get_body_cycle_numbers_df(
  .data,
  ph,
  pointval,
  .frame = frame,
  .point = point,
  .out = NULL,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{.data}{Data frame containing the midlines}

\item{ph}{Phase variable}

\item{pointval}{Specific point on the body to use to define the phase and
and cycle number}

\item{.frame, .point}{Columns identifying frames and points (defaults are \code{frame}
and \code{point})}

\item{.out}{Name of the output column. Needs to have one element specifying
the name for the cycle variable, or as a named list with an element named
\code{cycle}). Default is 'cycle'}

\item{overwrite}{TRUE or FALSE to overwrite existing columns, if present.}

\item{...}{Extra parameters to supply to \code{\link[=get_cycle_numbers]{get_cycle_numbers()}}}
}
\value{
A data frame containing a new column with the cycle numbers named
'cycle' or as specified in .out.
}
\description{
Extracts the phase for a specific point along the body (defined by \code{pointval})
and uses \code{\link[=get_cycle_numbers]{get_cycle_numbers()}} to get the cycle number. Joins the cycle
number data with the original data set so that cycle number is defined for
each point along the body
}
\concept{pipeline}
