% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{get_frequency}
\alias{get_frequency}
\title{Estimates the cycle frequency based on time and phase}
\usage{
get_frequency(
  t,
  ph,
  unwrap = FALSE,
  method = "deriv",
  mod = 2 * pi,
  check_reasonableness = TRUE
)
}
\arguments{
\item{t}{Time}

\item{ph}{Phase}

\item{unwrap}{(TRUE or FALSE) Unwrap the phase. Note that the
function will not work unless the phase is unwrapped, so you should only
set unwrap to FALSE if the phase has been unwrapped earlier.}

\item{method}{('deriv' or 'slope') Estimate the frequency based on}

\item{mod}{Modulus for the phase. Default is \code{2*pi}.}

\item{check_reasonableness}{Runs some simple checks to make sure the data
make sense. Checks to make sure phase is mostly increasing and warns if
it isn't.}
}
\value{
The frequency estimate, either as a vector the same size as \code{ph}
(for the 'deriv' algorithm) or as a single value (for the 'slope' algorithm)
}
\description{
Estimates frequency based on time and phase by one of two methods:
\itemize{
\item 'deriv' Takes the derivative of phase vs time to get frequency
\item 'slope' Fits a line to phase vs time and uses the slope as an estimate of
frequency.
}
}
\examples{
t <- seq(0, 3, by=0.1)
# example phase that has a frequency of exactly 2.3 Hz
ph <- 2*pi*(t * 2.3)

get_frequency(t, ph)
}
\concept{pipeline}
