% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{get_midline_center_df}
\alias{get_midline_center_df}
\title{Gets the center of a midline for many midlines in a data frame}
\usage{
get_midline_center_df(
  .data,
  arclen,
  x,
  y,
  mass,
  width,
  height,
  .out = NULL,
  .frame = frame,
  .point = point,
  excludepoints = c(),
  method = "mutate",
  overwrite = TRUE
)
}
\arguments{
\item{.data}{Data frame containing the midlines.}

\item{arclen}{The column containing the arc length. See \code{\link[=arclength]{arclength()}}}

\item{x, y}{Columns containing the x and y coordinates of the midline. There
should be N points.}

\item{mass}{(optional) Column containing the mass of each segment, with N-1
segments.}

\item{width}{(optional) Column containing the horizontal plane width of the
body at each midline point (N points)}

\item{height}{(optional) Column containing the dorso-ventral height of the
body at each midline point (N points)}

\item{.out}{Names of the output columns. Needs to have two elements specifying
the names for the x and y coordinates of center position. Or it can be a named
list containing at least some of the elements \code{xctr} and \code{yctr}. If the
return elements aren't in the named list, the defaults are 'xcom' and
'ycom')}

\item{.frame, .point}{Columns identifying frames and points (defaults are \code{frame}
and \code{point})}

\item{excludepoints}{Exclude these points when estimating center. Some points
(like the tip of the tail) have relatively little mass and are hard to track,
so can introduce errors.}

\item{method}{'mutate' or 'summarize'. If summarize, returns one center position
for each frame. If mutate, returns a same center position repeated for
each point in a frame.}

\item{overwrite}{TRUE or FALSE to overwrite existing columns, if present.}

\item{cutoff}{(optional) If this parameter is included, smooth the swimming
axis data with a low-pass filter with a cutoff at this frequency.}
}
\value{
A data frame containing the original variables along with
xcom, ycom (or names as specified in \code{.out}). The center of each midline
in each frame.
}
\description{
Estimates the center of a midline based on mass distribution, volume distribution,
or body width.
}
\details{
Given a mass distribution, it produces an estimates of the true center of mass.
If given the body width and height, it assumes that the body has an oval cross
section with varying width and height, and it estimates the volume distribution.
This method will give a good estimate of the center of mass if the body has
close to uniform density. If given just the width, it uses the width to
estimate a weight average centroid position.
}
\examples{
lampreydata |>
  group_by(frame) |>
  mutate(arclen = arclength(mxmm, mymm),
         width = interpolate_width(fishwidth$s, fishwidth$ammowidth, arclen)) |>
  get_midline_center_df(arclen, mxmm,mymm, width=width)
}
\concept{pipeline}
