% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{get_primary_swimming_axis}
\alias{get_primary_swimming_axis}
\title{Gets the main swimming axis from a midline}
\usage{
get_primary_swimming_axis(x, y, center = TRUE, na.rm = FALSE)
}
\arguments{
\item{x, y}{Coordinates of the midline}

\item{center}{(TRUE or FALSE) Subtract the mean from the x and y coordinates}

\item{na.rm}{(default FALSE) Remove NA points before computing the SVD}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{swimaxis_x}, \code{swimaxis_y} x and y components of the swimming axis vector
\item \code{swimaxis_xctr}, \code{swimaxis_yctr} Mean x and y values that were subtracted
before running the SVD
}
}
\description{
Computes the main swimming axis of a midline as a unit vector,
using the singular value
decomposition (\code{\link[=svd]{svd()}}). This only works well if the midlines are centered
around zero, so it optionally subtracts off the mean of x and y. For more
sophisticated centering algorithms, see \code{\link[=get_midline_center_df]{get_midline_center_df()}}.
}
\examples{
# run the algorithm across multiple midlines at different times
lampreydata |>
  group_by(t) |>
  summarize(swimaxis = get_primary_swimming_axis(mxmm, mymm)) |>
  unnest(swimaxis)
}
\seealso{
\code{\link[=get_midline_center_df]{get_midline_center_df()}}
}
