% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{get_volume}
\alias{get_volume}
\title{Gets the volume of segments of a cylindrical body with elliptical cross section}
\usage{
get_volume(arclen, width, height)
}
\arguments{
\item{arclen, width, height}{Arc length, width and height. Should have the same
units. N points}
}
\value{
Volume of each segment (N-1 values). Last value will be NA
}
\description{
Used for estimating the center of mass of a fish. If we know the width and
height profile, and we assume that the cross section is elliptical, then
we can estimate the volume of each segment as the volume of a truncated
elliptical cone.
}
\details{
The formula for such a cone is
V = pi s (w h + 1/2 dw h + 1/2 dh w + 1/3 dw dh)
where s is the length of the cone, w and h are the half width and height,
and dw and dh are the difference in width or height from
one end to the other (e.g., dw = w(s) - w(0) if w is a function of
s)
}
