% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{hilbert_phase}
\alias{hilbert_phase}
\title{Compute phase of an oscillation using the Hilbert transform}
\usage{
hilbert_phase(x, na.skip = TRUE, unwrap = TRUE, check_reasonableness = TRUE)
}
\arguments{
\item{check_reasonableness}{}
}
\value{
Phase
}
\description{
Compute phase of an oscillation using the Hilbert transform
}
\concept{pipeline}
