% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{interpolate_peak_location}
\alias{interpolate_peak_location}
\title{Interpolate the location of a peak based on three points}
\usage{
interpolate_peak_location(y, x = c(-1, 0, 1))
}
\arguments{
\item{x}{}
}
\description{
Uses a parabolic approximation to determine the location of a
peak from 3 points.
}
