% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{interpolate_points_df}
\alias{interpolate_points_df}
\title{Interpolates and smooths a 2D curve at new arc length}
\usage{
interpolate_points_df(
  .data,
  arclen,
  x,
  y,
  arclen_out = NULL,
  spar = 0.8,
  tailmethod = "extrapolate",
  .suffix = "_s",
  .out = NULL,
  overwrite = TRUE,
  .frame = frame,
  .point = point
)
}
\arguments{
\item{.data}{Data frame}

\item{arclen}{Name of the input arc length column in \code{.data}}

\item{x, y}{Name of the columns that contain the coordinates of points on the
curve}

\item{arclen_out}{Vector containing the new arc length}

\item{spar}{Smoothing parameter (ranges from 0 for no smoothing to 1 for
high smoothing; see \code{\link[=smooth.spline]{smooth.spline()}} for more details.)}

\item{tailmethod}{('keep', 'extrapolate', or 'NA') Methods to estimate the
position of the tail tip if the last value of \code{arclength_out} is longer
than the maximum arc length in the current frame.
\itemize{
\item 'keep' to keep the existing tail point, even if it is not at the
requested arc length
\item 'extrapolate' to extrapolate a tail tip position, assuming that the
curve continues straight
\item 'NA' to use replace the tail point with NA in this case.
}}

\item{.suffix}{(default = '_s') Suffix to append to the names of the
arclen, x, and y columns after smoothing and interpolation.}

\item{.out}{Names of the output columns. Defaults are
(arclen = 'arclen_s', xs = 'xs', ys = 'ys'). Overrides the \code{.suffix}
parameter if it is included.}

\item{overwrite}{TRUE or FALSE to overwrite existing columns}

\item{.frame}{Name of the frame variable in the data frame}

\item{.point}{Name of the point variable in the data frame}
}
\value{
A data frame with updated columns for the smoothed and iterpolated
arc length, x and y coordinates.
}
\description{
For a 2D curve with (x,y) coordinates parameterized by the arc length,
interpolate new (x,y) coordinates at new arc lengths. Smooth the input
data with a smoothing spline.
}
\details{
Operates on each frame (as defined in the \code{.frame} parameter) individually.
}
\concept{pipeline}
