% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{interpolate_points_frame}
\alias{interpolate_points_frame}
\title{Interpolates x and y points on a curve to different arc lengths}
\usage{
interpolate_points_frame(arclen, x, y, arclen_out, spar = 0.1, fill_gaps = 0)
}
\arguments{
\item{arclen}{Input arc length}

\item{x, y}{Coordinates of points on the curve}

\item{arclen_out}{New arc length}

\item{spar}{Smoothing parameter (ranges from 0 for no smoothing to 1 for
high smoothing; see \code{\link[=smooth.spline]{smooth.spline()}} for more details.)}

\item{fill_gaps}{Fill internal missing points of this size or smaller.
(0, default, means no filling; 1 means to fill single missing points)}
}
\value{
A tibble containing the new interpolated and smoothed x and y
coordinates as columns \code{xs} and \code{ys}
}
\description{
For a 2D curve with (x,y) coordinates parameterized by the arc length,
interpolate new (x,y) coordinates at new arc lengths. Smooth the input
data with a smoothing spline.
}
\examples{
# get one frame of lamprey midline data
df1 <- lampreydata[lampreydata$frame == 3, ]
# compute arc length along the midline
s <- with(df1, arclength(mxmm, mymm))
# define 20 evenly-spaced output arc lengths from head to tail
s_new <- seq(0, max(s, na.rm = TRUE), length.out = 20)
# interpolate and smooth the midline at the new arc lengths
with(df1, interpolate_points_frame(s, mxmm, mymm, s_new))
}
