% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_mass.R
\name{interpolate_width}
\alias{interpolate_width}
\title{Interpolates and scales fish body width}
\usage{
interpolate_width(arclen0, width0, arclen, scale_to_body_length = TRUE)
}
\arguments{
\item{arclen0}{Arc length for the width measurement. The first value should
be at the head and the last value should be at the tail tip.}

\item{width0}{Width measurement. Should have the same units as \code{arclen0}}

\item{arclen}{New arc length}

\item{scale_to_body_length}{TRUE or FALSE to scale the interpolated width
by multiplying by body length. This only works if \code{arclen} is in real units
(like cm) so that the last value in \code{arclen} is equal to the total length
of the fish.}
}
\value{
Width at the new values of arc length, scaled for the new length
}
\description{
Interpolates the width for a new arc length and scales it based on body length.
Assumes that the input width and arc length have the same units (they could be
in fractions of body length, cm, or pixels, as long as they are the same).
Once the width is estimated at the new arc length, scales it based on the new
maximum length.
}
\details{
Width here is defined as the distance from one side of the body to the other
(like a diameter), not from the center to a side (like a radius).
}
\concept{pipeline}
