% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_na}
\alias{skip_na}
\title{Skip NAs when running a function on a vector}
\usage{
skip_na(x, f, min.len = 1, ...)
}
\arguments{
\item{x}{A vector that may have NAs}

\item{f}{The function to run on the vector}

\item{...}{Other parameters to supply to the function}
}
\value{
A vector with the same length as \code{x} with NAs in the same places
}
\description{
\code{skip_na()} is a helper function related to \code{\link[=na.omit]{na.omit()}}. It runs a function
\code{f} on a vector that may contain NAs or NaNs, skipping all the NAs, and
returns the results as a vector of the same length as \code{x} with the NAs
in the same places.
}
\examples{
x <- c(1,2,3,NA,4,5,6,NA,7,8)
skip_na(x, cumsum)
# should return a vector the same length as x with NAs in position 4 and 8
}
