% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_points_df}
\alias{smooth_points_df}
\title{Smooths locations of points over time}
\usage{
smooth_points_df(
  .data,
  cols,
  spar,
  .out = NULL,
  .frame = frame,
  .point = point,
  fillgaps = 0
)
}
\arguments{
\item{.data}{Data frame containing the midlines.}

\item{cols}{Columns containing the components to be smoothed. Often these will
be the x and y coordinates of the midline.}

\item{.out}{Names of the output columns. Should either be a list with the same
number of elements as cols, or a glue specification as in \code{dplyr::across} for
the \code{.names} parameter. The default (NULL) means that the output columns will
have the same name as the original column with an 's' appended at the end.}

\item{.frame, .point}{Columns identifying frames and points (defaults are \code{frame}
and \code{point})}

\item{fillgaps}{Longest gap to interpolate over. default is 0, which means not
to fill gaps}
}
\value{
A data frame containing the smoothed points
}
\description{
Smooths columns specified in \code{cols} for each individual point over time, using
a smoothing spline. Optionally fills in gaps of less than \code{fillgaps} frames.
}
