% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xmucosusdata}
\alias{xmucosusdata}
\title{Prickleback tracking data}
\format{
A data frame with 711 rows and 27 columns. Columns follow the
pattern \verb{<keypoint>.x}, \verb{<keypoint>.y}, and \verb{<keypoint>.score} for each
tracked keypoint, plus \code{frame_idx}, \code{instance.score}, and \code{track}.
Keypoints are: \code{Snout}, \code{BP1}--\code{BP6}, and \code{Tail}.
}
\usage{
xmucosusdata
}
\description{
Tracking data for a a swimming rock prickleback, \emph{Xiphister mucosus}.
The data was tracked using Sleap (https://sleap.ai/) and comes out in the following format.
\code{frame_idx} is the frame number, and each point along the body is identified with the point
name and \code{.x}, \code{.y}, the coordinate, and \code{.score}, which is a measure of the estimated
accuracy of the point. All of the points together are also given a score (\code{instance.score}).
}
\keyword{datasets}
