% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zebrafish_shape}
\alias{zebrafish_shape}
\title{Zebrafish body shape}
\format{
A data frame with 20 rows and 3 columns:
\describe{
\item{s}{Arc length as a fraction of body length (0 = head, 1 = tail)}
\item{width}{Lateral body width as a fraction of body length}
\item{height}{Dorso-ventral body height as a fraction of body length}
}
}
\usage{
zebrafish_shape
}
\description{
Width and height profile of a zebrafish body as a function of fractional
arc length. Used for estimating volumes and centers of mass.
}
\keyword{datasets}
