% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zfish_goodframes}
\alias{zfish_goodframes}
\title{Zebrafish good frame ranges}
\format{
A data frame with 8 rows and 4 columns:
\describe{
\item{File}{File name of the corresponding trial in \code{zfishdata}}
\item{Start}{First usable frame index}
\item{End}{Last usable frame index}
\item{Block}{Block number within the trial}
}
}
\usage{
zfish_goodframes
}
\description{
Start and end frame indices for the usable portions of each zebrafish
swimming trial in \code{zfishdata}. Some trials contain multiple blocks of
good frames.
}
\keyword{datasets}
