% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{arclength}
\alias{arclength}
\title{Calculate arc length along a 2D curve}
\usage{
arclength(x, y, na.skip = FALSE)
}
\arguments{
\item{x, y}{Coordinates of the curve.}
}
\value{
Arc length along the curve.
}
\description{
Computes the straight-line distance between points on the curve and
then adds them up to get the arc length.
}
\examples{
# compute arc length in each frame from the lamprey data set
lampreydata |>
  group_by(frame) |>
  mutate(arclen = arclength(mxmm, mymm))

}
\concept{pipeline}
