% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishmechr-package.R
\docType{package}
\name{fishmechr-package}
\alias{fishmechr}
\alias{fishmechr-package}
\title{fishmechr: Computes Kinematic Parameters for Swimming}
\description{
Processes tracked points on a fish's body and uses them to estimate standard kinematic parameters such as tail beat amplitude and frequency, body wavelength and wavespeed. As part of this, it also estimates the location of the center of mass and the principal swimming axis. The techniques are described in detail in the main vignette and are published in the book chapter Hawkins, O.H., Di Santo, V., Tytell, Eric.D., 2025. "Biomechanics and energetics of swimming", in: Higham, T.E., Lauder, G.V. (Eds.), Integrative Fish Biomechanics, Fish Physiology. Academic Press. https://doi.org/10.1016/bs.fp.2025.06.003
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tytell.github.io/fishmechr/}
}

}
\author{
\strong{Maintainer}: Eric Tytell \email{eric.tytell@tufts.edu} (\href{https://orcid.org/0000-0002-6603-9448}{ORCID})

}
\keyword{internal}
