% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{interpolate_points_frame}
\alias{interpolate_points_frame}
\title{Interpolates x and y points on a curve to different arc lengths}
\usage{
interpolate_points_frame(arclen, x, y, arclen_out, spar = 0.1)
}
\arguments{
\item{arclen}{Input arc length}

\item{x, y}{Coordinates of points on the curve}

\item{arclen_out}{New arc length}

\item{spar}{Smoothing parameter (ranges from 0 for no smoothing to 1 for
high smoothing; see \code{\link[=smooth.spline]{smooth.spline()}} for more details.)}
}
\value{
A tibble containing the new interpolated and smoothed x and y
coordinates as columns \code{xs} and \code{ys}
}
\description{
For a 2D curve with (x,y) coordinates parameterized by the arc length,
interpolate new (x,y) coordinates at new arc lengths. Smooth the input
data with a smoothing spline.
}
