% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{peak_phase}
\alias{peak_phase}
\title{Compute phase of an oscillation by locating peaks and zero crossings}
\usage{
peak_phase(
  x,
  unwrap = TRUE,
  check_reasonableness = TRUE,
  check_ordering = TRUE,
  interpolate_peaks = TRUE,
  interpolate_zeros = TRUE,
  zero_mode = "midpoint",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Compute phase of an oscillation by locating peaks and zero crossings
}
\concept{pipeline}
