% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_kinematics.R
\name{pivot_kinematics_longer}
\alias{pivot_kinematics_longer}
\title{Pivots a kinematics dataset into long format}
\usage{
pivot_kinematics_longer(df, pointnames, point_to = "point", sep = "\\\\.")
}
\arguments{
\item{df}{The data frame}

\item{pointnames}{The names of the points, in order from head to tail}

\item{point_to}{The name of the column to put the point names in}

\item{sep}{The separator, as a regular expression (default = "\\.")}
}
\value{
The long data set
}
\description{
Pivots a kinematics dataset into long format
}
