% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_point}
\alias{smooth_point}
\title{Applies a smoothing spline to a data series, potentially with gaps}
\usage{
smooth_point(x, y, goodout = NULL, spar)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate, potentially with NAs}

\item{goodout}{Logical vector of where in the x coordinate to interpolate}

\item{spar}{Smoothing parameter (see \code{smooth.spline})}
}
\value{
The smoothed values
}
\description{
Builds a smoothing spline for y(x), where \code{y} may contain gaps (NAs).
Ignores the NAs. The uses the spline to interpolate values at the
same \code{x} coordinates, potentially filling in the gaps.
}
